requirements.txt
fastapi
uvicorn[standard]
requests
pandas
python-dotenv
oracledb
psycopg2-binary
